修改bug接口

+++ capacity_parent/capacity_front/src/main/java/com/zhidejiaoyu/student/business/feedback/service/impl/BugFeedBackServiceImpl.java
@@ -41,7 +41,7 @@ public class BugFeedBackServiceImpl extends BaseServiceImpl<BugFeedbackMapper, B
             feedback.setCreateTime(LocalDateTime.now());
             bugFeedbackMapper.insert(feedback);
         }
-        if (type != null) {
+        if (type != null && vocaId != null) {
             List<Integer> listType = getListType(type);
             studyCapacityMapper.deleteByTypeAndvocaId(listType, vocaId);
         }
